"Updated Kafka setup and updated data types for Event correlation IDs"
This commit involves updating the configuration setup for Kafka producers in "application.yml". With this change, the key and value serializers are defined under the namespace specific to the project instead of default Spring setup, making our configuration more clear and self-contained.
In the event handling code, the data type for Event's correlation Ids is changed from Map<CharSequence, CharSequence> to Map<String, String> to make the type interface more explicit and easier to work with.
Also, the output directory for Avro source generation was updated in the POM file. The change points the output to the build directory which is a more standard setup and will prevent potentially committing generated sources into the source code repository.